+++ plugins/backend/catalog-backend-module-okta/src/providers/OktaGroupEntityProvider.ts
@@ -32,6 +32,8 @@ import { AccountConfig } from '../types';
 import { groupEntityFromOktaGroup } from './groupEntityFromOktaGroup';
 import { getAccountConfig } from './accountConfig';
 import { assertError } from '@backstage/errors';
+import { getOktaGroups } from './getOktaGroups';
+import { getParentGroup } from './getParentGroup';
 
 /**
  * Provides entities from Okta Group service.
@@ -43,18 +45,32 @@ export class OktaGroupEntityProvider extends OktaEntityProvider {
   private readonly orgUrl: string;
   private readonly parentGroupField: string | undefined;
   private readonly customAttributesToAnnotationWhitelist: string[];
+  private hierarchyConfig: { parentKey: string; key?: string } | undefined;
 
   static fromConfig(
     config: Config,
     options: {
       logger: winston.Logger;
-      parentGroupField?: string;
       namingStrategy?: GroupNamingStrategies | GroupNamingStrategy;
       userNamingStrategy?: UserNamingStrategies | UserNamingStrategy;
       customAttributesToAnnotationWhitelist?: string[];
+      /*
+       * @deprecated, please use hierarchyConfig.parentKey
+       */
+      parentGroupField?: string;
+      hierarchyConfig?: {
+        parentKey: string;
+        key?: string;
+      };
     },
   ) {
     const accountConfig = getAccountConfig(config);
+
+    if (options.parentGroupField && !options.hierarchyConfig?.parentKey) {
+      options.hierarchyConfig = {
+        parentKey: options.parentGroupField,
+      };
+    }
 
     return new OktaGroupEntityProvider(accountConfig, options);
   }
@@ -63,14 +79,16 @@ export class OktaGroupEntityProvider extends OktaEntityProvider {
     accountConfig: AccountConfig,
     options: {
       logger: winston.Logger;
-      parentGroupField?: string;
       namingStrategy?: GroupNamingStrategies | GroupNamingStrategy;
       userNamingStrategy?: UserNamingStrategies | UserNamingStrategy;
       customAttributesToAnnotationWhitelist?: string[];
+      hierarchyConfig?: {
+        parentKey: string;
+        key?: string;
+      };
     },
   ) {
     super([accountConfig], options);
-    this.parentGroupField = options.parentGroupField;
     this.namingStrategy = groupNamingStrategyFactory(options.namingStrategy);
     this.userNamingStrategy = userNamingStrategyFactory(
       options.userNamingStrategy,
@@ -81,6 +99,7 @@ export class OktaGroupEntityProvider extends OktaEntityProvider {
       options.customAttributesToAnnotationWhitelist
         ? options.customAttributesToAnnotationWhitelist
         : [];
+    this.hierarchyConfig = options.hierarchyConfig;
   }
 
   getProviderName(): string {
@@ -99,15 +118,47 @@ export class OktaGroupEntityProvider extends OktaEntityProvider {
 
     const defaultAnnotations = await this.buildDefaultAnnotations();
 
-    await client.listGroups({ search: this.groupFilter }).each(async group => {
-      const members: string[] = [];
-      await group.listUsers().each(user => {
+    const oktaGroups = await getOktaGroups({
+      client,
+      groupFilter: this.groupFilter,
+      groupNamingStrategy: this.namingStrategy,
+      key: this.hierarchyConfig?.key,
+      logger: this.logger,
+    });
+
+    await Promise.allSettled(
+      Object.entries(oktaGroups).map(async ([_, group]) => {
+        const members: string[] = [];
+        await group.listUsers().each(user => {
+          try {
+            const userName = this.userNamingStrategy(user);
+            members.push(userName);
+          } catch (e) {
+            assertError(e);
+            this.logger.warn(`failed to add user to group: ${e.message}`);
+          }
+        });
+
+        const parentGroup = getParentGroup({
+          parentKey: this.hierarchyConfig?.parentKey,
+          group,
+          oktaGroups,
+        });
+
         try {
-          const userName = this.userNamingStrategy(user);
-          members.push(userName);
+          const groupEntity = groupEntityFromOktaGroup(
+            group,
+            this.namingStrategy,
+            parentGroup,
+            {
+              annotations: defaultAnnotations,
+              members,
+            },
+          );
+          groupResources.push(groupEntity);
         } catch (e) {
           assertError(e);
-          this.logger.warn(`failed to add user to group: ${e.message}`);
+          this.logger.warn(`failed to add group: ${e.message}`);
         }
       });
       const profileAnnotations: Record<string, string> = {};