+++ plugins/backend/catalog-backend-module-okta/src/providers/OktaOrgEntityProvider.ts
@@ -33,6 +33,9 @@ import { AccountConfig } from '../types';
 import { groupEntityFromOktaGroup } from './groupEntityFromOktaGroup';
 import { getAccountConfig } from './accountConfig';
 import { assertError } from '@backstage/errors';
+import { getOktaGroups } from './getOktaGroups';
+import { getParentGroup } from './getParentGroup';
+import { GroupTree } from './GroupTree';
 
 /**
  * Provides entities from Okta Org service.
@@ -40,21 +43,37 @@ import { assertError } from '@backstage/errors';
 export class OktaOrgEntityProvider extends OktaEntityProvider {
   private readonly groupNamingStrategy: GroupNamingStrategy;
   private readonly userNamingStrategy: UserNamingStrategy;
-  private readonly parentGroupField: string | undefined;
+  private readonly includeEmptyGroups: boolean;
+  private readonly hierarchyConfig:
+    | { parentKey: string; key?: string }
+    | undefined;
 
   static fromConfig(
     config: Config,
     options: {
       logger: winston.Logger;
-      parentGroupField?: string;
       groupNamingStrategy?: GroupNamingStrategies | GroupNamingStrategy;
       userNamingStrategy?: UserNamingStrategies | UserNamingStrategy;
+      includeEmptyGroups?: boolean;
+      /*
+       * @deprecated, please use hierarchyConfig.parentKey
+       */
+      parentGroupField?: string;
+      hierarchyConfig?: {
+        parentKey: string;
+        key?: string;
+      };
     },
   ) {
     const oktaConfigs = config
       .getOptionalConfigArray('catalog.providers.okta')
       ?.map(getAccountConfig);
 
+    if (options.parentGroupField && !options.hierarchyConfig?.parentKey) {
+      options.hierarchyConfig = {
+        parentKey: `profile.${options.parentGroupField}`,
+      };
+    }
     return new OktaOrgEntityProvider(oktaConfigs || [], options);
   }
 
@@ -62,19 +81,24 @@ export class OktaOrgEntityProvider extends OktaEntityProvider {
     accountConfigs: AccountConfig[],
     options: {
       logger: winston.Logger;
-      parentGroupField?: string;
       groupNamingStrategy?: GroupNamingStrategies | GroupNamingStrategy;
       userNamingStrategy?: UserNamingStrategies | UserNamingStrategy;
+      includeEmptyGroups?: boolean;
+      hierarchyConfig?: {
+        parentKey: string;
+        key?: string;
+      };
     },
   ) {
     super(accountConfigs, options);
-    this.parentGroupField = options.parentGroupField;
     this.groupNamingStrategy = groupNamingStrategyFactory(
       options.groupNamingStrategy,
     );
     this.userNamingStrategy = userNamingStrategyFactory(
       options.userNamingStrategy,
     );
+    this.includeEmptyGroups = !!options.includeEmptyGroups;
+    this.hierarchyConfig = options.hierarchyConfig;
   }
 
   getProviderName(): string {
@@ -87,7 +111,8 @@ export class OktaOrgEntityProvider extends OktaEntityProvider {
     }
 
     this.logger.info('Providing user and group resources from okta');
-    const resources: (GroupEntity | UserEntity)[] = [];
+    let groupResources: GroupEntity[] = [];
+    const userResources: Record<string, UserEntity> = {};
     let providedUserCount = 0;
     let providedGroupCount = 0;
 
@@ -101,58 +126,89 @@ export class OktaOrgEntityProvider extends OktaEntityProvider {
         const defaultAnnotations = await this.buildDefaultAnnotations();
 
         await client.listUsers({ search: account.userFilter }).each(user => {
-          resources.push(
-            userEntityFromOktaUser(user, this.userNamingStrategy, {
-              annotations: defaultAnnotations,
-            }),
-          );
+          try {
+            const userName = this.userNamingStrategy(user);
+            userResources[userName] = userEntityFromOktaUser(
+              user,
+              this.userNamingStrategy,
+              {
+                annotations: defaultAnnotations,
+              },
+            );
+          } catch (e: unknown) {
+            assertError(e);
+            this.logger.warn(`Failed to add user: ${e.message}`);
+          }
         });
 
-        providedUserCount = resources.length;
+        providedUserCount = Object.values(userResources).length;
 
-        await client
-          .listGroups({ search: account.groupFilter })
-          .each(async group => {
+        const oktaGroups = await getOktaGroups({
+          client,
+          groupFilter: account.groupFilter,
+          key: this.hierarchyConfig?.key,
+          groupNamingStrategy: this.groupNamingStrategy,
+          logger: this.logger,
+        });
+
+        await Promise.allSettled(
+          Object.entries(oktaGroups).map(async ([_, group]) => {
             const members: string[] = [];
             await group.listUsers().each(user => {
               try {
                 const userName = this.userNamingStrategy(user);
-                members.push(userName);
+                if (userResources[userName]) {
+                  members.push(userName);
+                }
               } catch (e: unknown) {
                 assertError(e);
                 this.logger.warn(`failed to add user to group: ${e.message}`);
               }
             });
 
+            const parentGroup = getParentGroup({
+              parentKey: this.hierarchyConfig?.parentKey,
+              group,
+              oktaGroups,
+            });
+
             try {
               const groupEntity = groupEntityFromOktaGroup(
                 group,
                 this.groupNamingStrategy,
+                parentGroup,
                 {
                   annotations: defaultAnnotations,
                   members,
-                  parentGroupField: this.parentGroupField,
                 },
               );
-              if (members.length > 0) {
-                resources.push(groupEntity);
-              }
+              groupResources.push(groupEntity);
             } catch (e: unknown) {
               assertError(e);
               this.logger.warn(`failed to add group: ${e.message}`);
             }
-          });
+          }),
+        );
       }),
     );
 
-    providedGroupCount = resources.length - providedUserCount;
+    if (!this.includeEmptyGroups) {
+      groupResources = new GroupTree(groupResources).getGroups({
+        pruneEmptyMembers: true,
+      });
+    }
+
+    providedGroupCount = groupResources.length;
     await this.connection.applyMutation({
       type: 'full',
-      entities: resources.map(entity => ({
-        entity,
-        locationKey: this.getProviderName(),
-      })),
+      entities: [...Object.values(userResources), ...groupResources].map(
+        entity => ({
+          entity,
+          locationKey: this.getProviderName(),
+        }),
+      ),
     });
+
     this.logger.info(
       `Finished providing ${providedUserCount} user and ${providedGroupCount} group resources from okta`,
     );