+++ plugins/backend/catalog-backend-module-okta/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @roadiehq/catalog-backend-module-okta
 
+## 0.7.0
+
+### Minor Changes
+
+- 46cc9fc7: Add smarter pruning of empty groups. Previously, parent groups of groups with members were being pruned. This change includes parent groups if their decendants have members.
+
+## 0.6.1
+
+### Patch Changes
+
+- 7c561d79: Uses imports from the `@backstage/plugin-catalog-node` package
+
+## 0.6.0
+
+### Minor Changes
+
+- 7baa7791: Rewrite the configuration and logic for creating a hierarchy of groups. Now it does not insist on the group name matching the parent key.
+
+## 0.5.7
+
+### Patch Changes
+
+- e331d3a1: Bump to backstage version 1.12.1
+
+## 0.5.6
+
+### Patch Changes
+
+- cc0f1489: Allow the okta provider to include groups that have no members.
+
 ## 0.5.5
 
 ### Patch Changes