+++ plugins/frontend/backstage-plugin-argo-cd/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-argo-cd
 
+## 2.2.10
+
+### Patch Changes
+
+- e331d3a1: Bump to backstage version 1.12.1
+
 ## 2.2.9
 
 ### Patch Changes