+++ plugins/frontend/backstage-plugin-buildkite/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-buildkite
 
+## 2.1.8
+
+### Patch Changes
+
+- e331d3a1: Bump to backstage version 1.12.1
+
 ## 2.1.7
 
 ### Patch Changes