+++ plugins/frontend/backstage-plugin-github-pull-requests/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-github-pull-requests
 
+## 2.5.6
+
+### Patch Changes
+
+- e331d3a1: Bump to backstage version 1.12.1
+
 ## 2.5.5
 
 ### Patch Changes