+++ plugins/frontend/backstage-plugin-iframe/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-iframe
 
+## 1.3.2
+
+### Patch Changes
+
+- e331d3a1: Bump to backstage version 1.12.1
+
 ## 1.3.1
 
 ### Patch Changes