+++ plugins/frontend/backstage-plugin-security-insights/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-security-insights
 
+## 2.1.11
+
+### Patch Changes
+
+- e331d3a1: Bump to backstage version 1.12.1
+
 ## 2.1.10
 
 ### Patch Changes